refactor(eslint-config): remove next config
