+++ packages/eslint-config/README.md
@@ -1,3 +1,3 @@
-# `@turbo/eslint-config`
+# `@repo/eslint-config`
 
 Collection of internal eslint configurations.
