+++ packages/eslint-config/base.js
@@ -2,7 +2,6 @@ import js from "@eslint/js";
 import eslintConfigPrettier from "eslint-config-prettier";
 import turboPlugin from "eslint-plugin-turbo";
 import tseslint from "typescript-eslint";
-import onlyWarn from "eslint-plugin-only-warn";
 
 /**
  * A shared ESLint configuration for the repository.
@@ -22,11 +21,6 @@ export const config = [
 		},
 	},
 	{
-		plugins: {
-			onlyWarn,
-		},
-	},
-	{
-		ignores: ["dist/**"],
+		ignores: ["dist/**", "coverage/**"],
 	},
 ];
