+++ packages/eslint-config/package.json
@@ -5,15 +5,12 @@
 	"private": true,
 	"exports": {
 		"./base": "./base.js",
-		"./next-js": "./next.js",
-		"./react-internal": "./react-internal.js"
+		"./react": "./react.js"
 	},
 	"devDependencies": {
 		"@eslint/js": "^9.39.1",
-		"@next/eslint-plugin-next": "^16.2.0",
 		"eslint": "^9.39.1",
 		"eslint-config-prettier": "^10.1.1",
-		"eslint-plugin-only-warn": "^1.1.0",
 		"eslint-plugin-react": "^7.37.5",
 		"eslint-plugin-react-hooks": "^5.2.0",
 		"eslint-plugin-turbo": "^2.7.1",
