+++ packages/eslint-config/react.js
@@ -1,0 +1,32 @@
+import pluginReact from "eslint-plugin-react";
+import pluginReactHooks from "eslint-plugin-react-hooks";
+import globals from "globals";
+import { config as baseConfig } from "./base.js";
+
+/**
+ * A shared ESLint configuration for React code.
+ *
+ * @type {import("eslint").Linter.Config[]}
+ */
+export const config = [
+	...baseConfig,
+	{
+		...pluginReact.configs.flat.recommended,
+		languageOptions: {
+			...pluginReact.configs.flat.recommended.languageOptions,
+			globals: {
+				...globals.browser,
+			},
+		},
+	},
+	pluginReact.configs.flat["jsx-runtime"],
+	{
+		plugins: {
+			"react-hooks": pluginReactHooks,
+		},
+		settings: { react: { version: "detect" } },
+		rules: {
+			...pluginReactHooks.configs.recommended.rules,
+		},
+	},
+];
